+++ src/main/java/com/rackspace/salus/zw/handler/ActiveZoneEventProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.zone_watcher.handler;
+package com.rackspace.salus.zw.handler;
 
 import static com.rackspace.salus.telemetry.etcd.types.Keys.PTN_ZONE_ACTIVE;
 import static com.rackspace.salus.telemetry.etcd.types.Keys.TRACKING_KEY_ZONE_ACTIVE;
@@ -25,7 +25,7 @@ import com.coreos.jetcd.data.ByteSequence;
 import com.coreos.jetcd.watch.WatchEvent;
 import com.coreos.jetcd.watch.WatchResponse;
 import com.rackspace.salus.telemetry.etcd.services.ZoneStorage;
-import com.rackspace.salus.zone_watcher.services.ZoneStorageListener;
+import com.rackspace.salus.zw.services.ZoneStorageListener;
 import com.rackspace.salus.telemetry.etcd.types.ResolvedZone;
 import java.util.Optional;
 import java.util.regex.Matcher;
