+++ src/test/java/com/rackspace/salus/zw/services/ZoneWatchingServiceTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.zone_watcher.services;
+package com.rackspace.salus.zw.services;
 
 import static com.rackspace.salus.telemetry.etcd.types.ResolvedZone.createPrivateZone;
 import static org.hamcrest.Matchers.equalTo;
